refactor: mod python folder structure to match style of aws-rekogition repo
