+++ .pylintrc
@@ -1,5 +1,4 @@
 [MASTER]
-init-hook='import sys; print(sys.executable); sys.path.append("/absolute/path/to/api/terraform/python/layer_genai/openai_utils")'
 ignore-paths =
   api/terraform/python/lambda_openai_v2/lambda_dist_pkg,
   api/terraform/python/lambda_openai/lambda_dist_pkg,
