+++ api/terraform/apigateway_endpoint_passthrough_v2.tf
@@ -44,7 +44,7 @@ resource "aws_api_gateway_integration" "passthrough_v2_post" {
   http_method             = aws_api_gateway_method.passthrough_v2_post.http_method
   integration_http_method = "POST"
   type                    = "AWS"
-  uri                     = aws_lambda_function.lambda_openai_v2_.invoke_arn
+  uri                     = aws_lambda_function.lambda_openai_v2.invoke_arn
   credentials             = aws_iam_role.apigateway.arn
   passthrough_behavior    = "WHEN_NO_TEMPLATES"
   depends_on              = [aws_api_gateway_method.passthrough_v2_post]
