+++ api/terraform/apigateway_tests.tf
@@ -1,3 +1,19 @@
+locals {
+  template_file_test_200 = templatefile("${path.module}/templates/test_200.json", {})
+  template_file_test_400 = templatefile("${path.module}/templates/test_400.json", {})
+  template_file_test_500 = jsonencode({
+    "isBase64Encoded" : false,
+    "statusCode" = 500
+    "body" : {
+      "error" : "Internal Server Error",
+      "message" : "TEST 500 RESPONSE."
+    }
+  })
+
+  template_file_test_504 = jsonencode({
+    "statusCode" = 504
+  })
+}
 resource "aws_api_gateway_resource" "tests" {
   path_part   = "tests"
   parent_id   = aws_api_gateway_rest_api.openai.root_resource_id
@@ -7,10 +23,6 @@ resource "aws_api_gateway_resource" "tests" {
 ###############################################################################
 # Test 200
 ###############################################################################
-
-data "template_file" "test_200" {
-  template = file("${path.module}/templates/test_200.json")
-}
 
 resource "aws_api_gateway_resource" "test_200" {
   path_part   = "test_200"
@@ -33,7 +45,7 @@ resource "aws_api_gateway_integration" "test_200" {
   integration_http_method = "POST"
   type                    = "MOCK"
   request_templates = {
-    "application/json" = data.template_file.test_200.rendered
+    "application/json" = local.template_file_test_200
   }
   depends_on = [aws_api_gateway_method.test_200]
   lifecycle {
@@ -73,7 +85,7 @@ resource "aws_api_gateway_integration_response" "test_200" {
     "method.response.header.Access-Control-Allow-Origin"  = "'*'"
   }
   response_templates = {
-    "application/json" = data.template_file.test_200.rendered
+    "application/json" = local.template_file_test_200
   }
 
   depends_on = [
@@ -85,9 +97,6 @@ resource "aws_api_gateway_integration_response" "test_200" {
 ###############################################################################
 # Test 400
 ###############################################################################
-data "template_file" "test_400" {
-  template = file("${path.module}/templates/test_400.json")
-}
 
 resource "aws_api_gateway_resource" "test_400" {
   path_part   = "test_400"
@@ -110,7 +119,7 @@ resource "aws_api_gateway_integration" "test_400" {
   integration_http_method = "POST"
   type                    = "MOCK"
   request_templates = {
-    "application/json" = data.template_file.test_400.rendered
+    "application/json" = local.template_file_test_400
   }
   depends_on = [aws_api_gateway_method.test_400]
   lifecycle {
@@ -150,7 +159,7 @@ resource "aws_api_gateway_integration_response" "test_400" {
     "method.response.header.Access-Control-Allow-Origin"  = "'*'"
   }
   response_templates = {
-    "application/json" = data.template_file.test_400.rendered
+    "application/json" = local.template_file_test_400
   }
 
   depends_on = [
@@ -162,16 +171,6 @@ resource "aws_api_gateway_integration_response" "test_400" {
 ###############################################################################
 # Test 500
 ###############################################################################
-data "template_file" "test_500" {
-  template = jsonencode({
-    "isBase64Encoded" : false,
-    "statusCode" = 500
-    "body" : {
-      "error" : "Internal Server Error",
-      "message" : "TEST 500 RESPONSE."
-    }
-  })
-}
 
 resource "aws_api_gateway_resource" "test_500" {
   path_part   = "test_500"
@@ -194,7 +193,7 @@ resource "aws_api_gateway_integration" "test_500" {
   integration_http_method = "POST"
   type                    = "MOCK"
   request_templates = {
-    "application/json" = data.template_file.test_500.rendered
+    "application/json" = local.template_file_test_500
   }
   depends_on = [aws_api_gateway_method.test_500]
   lifecycle {
@@ -234,7 +233,7 @@ resource "aws_api_gateway_integration_response" "test_500" {
     "method.response.header.Access-Control-Allow-Origin"  = "'*'"
   }
   response_templates = {
-    "application/json" = data.template_file.test_500.rendered
+    "application/json" = local.template_file_test_500
   }
 
   depends_on = [
@@ -245,11 +244,6 @@ resource "aws_api_gateway_integration_response" "test_500" {
 ###############################################################################
 # Test 504
 ###############################################################################
-data "template_file" "test_504" {
-  template = jsonencode({
-    "statusCode" = 504
-  })
-}
 
 resource "aws_api_gateway_resource" "test_504" {
   path_part   = "test_504"
@@ -272,7 +266,7 @@ resource "aws_api_gateway_integration" "test_504" {
   integration_http_method = "POST"
   type                    = "MOCK"
   request_templates = {
-    "application/json" = data.template_file.test_504.rendered
+    "application/json" = local.template_file_test_504
   }
   depends_on = [aws_api_gateway_method.test_504]
   lifecycle {
