+++ api/terraform/endpoint/api_gateway_endpoint.tf
@@ -4,6 +4,19 @@
 #
 # see: https://mrponath.medium.com/terraform-and-aws-api-gateway-a137ee48a8ac
 ###############################################################################
+locals {
+  openai_integration_template = templatefile(
+    "${path.module}/mapping-templates/openai-integration.tpl",
+    {
+      mapping_model               = var.mapping_model
+      mapping_role_system_content = var.mapping_role_system_content
+      mapping_end_point           = var.mapping_end_point
+      mapping_temperature         = var.mapping_temperature
+      mapping_max_tokens          = var.mapping_max_tokens
+    }
+  )
+
+}
 resource "aws_api_gateway_resource" "endpoint" {
   path_part   = var.path_part
   parent_id   = var.aws_api_gateway_rest_api_parent_id
@@ -21,16 +34,6 @@ resource "aws_api_gateway_method" "post" {
   api_key_required = "true"
 }
 
-data "template_file" "openai_integration" {
-  template = file("${path.module}/mapping-templates/openai-integration.tpl")
-  vars = {
-    mapping_model               = var.mapping_model
-    mapping_role_system_content = var.mapping_role_system_content
-    mapping_end_point           = var.mapping_end_point
-    mapping_temperature         = var.mapping_temperature
-    mapping_max_tokens          = var.mapping_max_tokens
-  }
-}
 
 resource "aws_api_gateway_integration" "post" {
   rest_api_id             = var.aws_api_gateway_rest_api_id
@@ -43,7 +46,7 @@ resource "aws_api_gateway_integration" "post" {
   # https://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-mapping-template-reference.html#input-variable-reference
   # https://goessner.net/articles/JsonPath/
   request_templates = {
-    "application/json" = data.template_file.openai_integration.rendered
+    "application/json" = local.openai_integration_template
   }
   passthrough_behavior = "WHEN_NO_TEMPLATES"
   depends_on           = [aws_api_gateway_method.post]
