+++ api/terraform/python/layer_genai/README.md
@@ -1,0 +1,3 @@
+# AWS Lambda Layer for OpenAI/Langchain Lambdas
+
+This layer contains the combined pip requirements for both lambda_langchain and lambda_openai.
