--- api/terraform/python/layer_genai/openai_utils/__version__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-__version__ = '0.6.5'
-
