+++ api/terraform/python/layer_genai/requirements.txt
@@ -8,7 +8,7 @@
 #             requirements in the zip package for this layer.
 # -----------------------------------------------------------------------------
 python-dotenv==1.0.0
-openai==1.3.8
+openai==1.4.0
 langchain==0.0.348
-langchain-experimental==0.0.45
+langchain-experimental
 pinecone-client==2.2.4
