+++ api/terraform/python/openai_api/README.md
@@ -1,0 +1,5 @@
+# AWS Lambdas for API Gateway
+
+These are the Lambdas that are called from API Gateway endpoint resources.
+
+Note that **version**.py is maintained by Github Actions automations.
