+++ api/terraform/python/openai_api/__version__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# Managed via automated CI/CD in .github/workflows/semanticVersionBump.yml.
+__version__ = "0.6.5-next.1"
