+++ api/terraform/python/openai_api/common/.gitignore
