+++ api/terraform/python/openai_api/common/tests/mock_data/example_request.json
