+++ api/terraform/python/openai_api/lambda_langchain/.gitignore
