+++ api/terraform/python/openai_api/lambda_langchain/env.sh
