+++ api/terraform/python/openai_api/lambda_langchain/tests/events/langchain.response.json
