+++ api/terraform/python/openai_api/lambda_langchain/tests/events/test_01.request.json
