+++ api/terraform/python/openai_api/lambda_openai_v2/env.sh
