+++ api/terraform/python/openai_api/lambda_openai_v2/tests/events/test_01.response.json
