+++ api/terraform/terraform.tf
@@ -21,7 +21,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 5.8"
+      version = "~> 5.31"
     }
     null = {
       source  = "hashicorp/null"
