+++ requirements.txt
@@ -31,10 +31,6 @@ codespell==2.2.6
 # production
 # ------------
 python-dotenv==1.0.0
-openai==1.3.7
-langchain==0.0.345
+openai==1.4.0
 langchain-experimental==0.0.43
 pinecone-client==2.2.4
-
-# local packages
--e ./api/terraform/python/layer_genai/openai_utils
