+++ tox.ini
@@ -45,8 +45,4 @@ commands = bump2version {posargs}
 [testenv:pylint]
 deps = pylint
 commands =
-    pylint api/terraform/python/lambda_langchain
-    pylint api/terraform/python/lambda_openai
-    pylint api/terraform/python/lambda_openai_v2
-    pylint api/terraform/python/layer_genai
-    pylint api/terraform/python/layer_genai/openai_utils
+    pylint api/terraform/python/
